+++ src/index.ts
@@ -8,6 +8,8 @@ export * from "./lib/types/LabelValueOption";
 export * from "./lib/DatePickerInput";
 export * from "./lib/helpers/dateUtils";
 
+export { useInternalFormContext } from "./lib/context/InternalFormContext";
+
 // for some unknown reason, SSR only works if the consumer
 // imports useFormContext from our package
 export { useFormContext } from "react-hook-form";